I added express and templating language ejs.
I thought we won't need any templating language, but I think it may
be useful for stuff like displaying user name and initial resources
and stuff. Unless we want to load everything through websockets. Anyway
EJS is much closer to html than Jade, and we don't need all of that
shit that Jade does. Check it out, fic it a little bit. I think it's
pretty cool and does all the routing and shit for us. There even is a
session manager but I don't knwo how to use it yet.