Implemented a very simple messaging system.
Users can message each other, and all the messages to them show
up in their inbox (/inbox).

A problem I encountered though, is with Mongoose. It does not allow
to do things such as

to: UserSchema
from: UserSchema

for some random reason..... The way around it is to declare anything like that
as ObjectId (a unique thing in Mongo per collection), and then fetch stuff out
using ObjectId.

If anyone can be bothered to do it, how am I able to asynchronously do a forEach
on a result that returns multiple object, and modify each object, and only then
continue with my next callback ?
I would need something like this to map the ObjectId's to username before sending
off the data to the client side.