Dynamic message notification is on the way.
It currently works, but in a really crappy way.
Log in as 2 different users, go to inbox in one of them, but not in the other. Send a message to the other, you should get a notification at
the other end.

The problem is that anytime you change a page, socket.io
(on the server), will issue us with a new 'client' object, which I use
to communicate and send data with. Hence, someone find a solution.

Should I just keep an array of 'clients' for every username, or maybe
restrict the user to only have 1 window ever open and not open
any new windows with our game.

Additionally, maybe we should it, so that instead of doing a page
reload to get to /inbox (or wherever),  maybe we should just use AJAX and
dynamically update a container div that would hold all the content and whatever. That would mean I would never get disconnected
from socket.io, which would be quite handy.

Any progress on anything else ? Map editor / css ?