Added unselect using escape.
For now it's a part of player, but I think it will be better to make
it a global thing that unselects whatever is selected.

Also changed the name of setScale to setBitmapScale in planet, so we
have consistency in case wee want to make a global thing.

And added some missing semicolons.